refactor(#1526): extract recall result formatter (seam 8)
Extract the recall result formatting + identity continuity section assembly
from orchestrator.ts into a dedicated RecallResultFormatter coordinator:

- formatQmdResults (memory handles + epistemic hedge rendering)
- formatObjectiveStateResults
- formatCausalTrajectoryResults
- formatTrustZoneResults
- formatHarmonicRetrievalResults
- formatWorkProductResults
- formatVerifiedEpisodeResults
- formatVerifiedSemanticRuleResults
- summarizeIdentityText, formatOpenIncidentLine, trimIdentitySection
- buildIdentityContinuitySection (identity section assembly)
- hasIdentityRecoveryIntent + resolveEffectiveIdentityInjectionMode helpers

Behavior-preserving move: the orchestrator keeps thin delegating private
methods so call sites and tests that exercise the private API are unaffected.

orchestrator.ts: 19737 → 19454 LOC (-283)

Chokepoints used: n/a (refactor only, no new surfaces).
